# swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: "provaEsame"
  description: Managing and monitoring assignments and exams
  termsOfService: http://swagger.io/terms/
  contact:
    name: Enrico Bontorin
    email: enricobontorin@gmail.com
    url: http://exams.jointresearch.net
  license:
    name: MIT
    url: http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT
host: examse2.herokuapp.com
basePath: /
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /assignment:
    get:
      description: Retrieves assigments
      operationId: getAssignments
      responses:
        200:
          description: list of assignments
          schema:
            type: array
            items:
              $ref: '#/definitions/Assignment'
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      description: Uploads a new assignment.
      operationId: addAssignment
      parameters:
        - name: assignment
          in: body
          description: new assignment
          required: true
          schema:
            $ref: '#/definitions/NewAssignment'
      responses:
        200:
          description: assignment created response
          schema:
            $ref: '#/definitions/Assignment'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /assignment/{id}:
    get:
      description: Returns an assignment
      operationId: findbyassignment id
      parameters:
        - name: id
          in: path
          description: ID of assignment to fetch
          required: true
          type: string
          x-example: 'some-example'
      responses:
        200:
          description: assignment
          schema:
            items:
              $ref: '#/definitions/Assignment'
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      description: Updates an existing assignment, or create a new one with ID.
      operationId: updateAssignment
      parameters:
        - name: id
          in: path
          description: ID of assignment to edit or create
          required: true
          type: string
          x-example: 'some-example'
        - name: assignment
          in: body
          description: assignment
          required: true
          schema:
            $ref: '#/definitions/Assignment'
      responses:
        200:
          description: assignment put
          schema:
            items:
              $ref: '#/definitions/Assignment'
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      description: deletes a single assignment based on the ID supplied
      operationId: deleteAssignment
      parameters:
        - name: id
          in: path
          description: ID of assignment to delete
          required: true
          type: string
          x-example: 'erwtre34-36fws'
      responses:
        200:
          description: message
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
  Assignment:
    allOf:
      - $ref: '#/definitions/NewAssignment'
      - required:
        - assignmentID
        properties:
          assignmentID:
            type: string
          ipAddress:
            type: string
            format: ipv6
          dateUpdated:
            type: string
            format: date-time
  NewAssignment:
    required:
      - workerID
      - taskID
      - assignmentResult
    properties:
      workerID:
        type: string
      taskID:
        type: string
      assignmentResult:
        type: string
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string